requirements.txt
django==3.2
psycopg2-binary==2.8.6
djangorestframework==3.13.1
